Add namespace to DIC tests
